fix(compiler): move instruction compiler utilities
Move instruction-compiler-only helpers under instructionCompilers/utils, update imports directly without compatibility exports, and remove the stale memoryIdentifier utility.
