+++ packages/compiler/src/instructionCompilers/utils/createResultBlockState.ts
@@ -2,7 +2,7 @@ import type { BlockResultTypes, BlockStack, BlockType, FunctionTypeRegistry } fr
 
 import type { WasmTypeValue } from '@8f4e/compiler-wasm-utils';
 import { signedLEB128, WASM_TYPE_F32, WASM_TYPE_I32, WASM_TYPE_VOID } from '@8f4e/compiler-wasm-utils';
-import { getOrRegisterFunctionType } from '../../utils/functionTypeRegistry';
+import { getOrRegisterFunctionType } from './functionTypeRegistry';
 
 type ResultBlockType = typeof BlockType.BLOCK | typeof BlockType.CONDITION;
