--- packages/compiler/src/utils/memoryIdentifier.test.ts
@@ -1,22 +1,0 @@
-import type { MemoryMap } from '@8f4e/compiler-spec';
-import { describe, expect, it } from 'vitest';
-import { isMemoryIdentifier } from './memoryIdentifier';
-
-describe('memoryIdentifier utilities', () => {
-	const mockMemory: MemoryMap = {
-		foo: {} as unknown as MemoryMap[string],
-		bar: {} as unknown as MemoryMap[string],
-	};
-
-	describe('isMemoryIdentifier', () => {
-		it('returns true for existing memory identifiers', () => {
-			expect(isMemoryIdentifier(mockMemory, 'foo')).toBe(true);
-			expect(isMemoryIdentifier(mockMemory, 'bar')).toBe(true);
-		});
-
-		it('returns false for non-existing memory identifiers', () => {
-			expect(isMemoryIdentifier(mockMemory, 'baz')).toBe(false);
-			expect(isMemoryIdentifier(mockMemory, '')).toBe(false);
-		});
-	});
-});
